api/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-openai
chromadb
psycopg2-binary